gmail/requirements.txt
beautifulsoup4
google-auth-oauthlib==1.0.0
google-api-python-client==2.81.0
flask
